+++ mainapp/src/Log.ts
@@ -1,10 +1,28 @@
-type Log = { user: string; tool: string; source: string; timestamp: string };
+type Log = { user: string; tool: string; source: string; action: string; timestamp: string };
 
 const endpoint = "/api/v1/logs";
-async function getLogs(): Promise<Log[]> {
+async function getLogFiles(): Promise<Log[]> {
     const response = await fetch(endpoint);
     const json = (await response.json()) as Log[];
     return json;
 }
+async function getLogs(file): Promise<Log[]> {
+    const response = await fetch(endpoint + "/" + file);
+    const json = (await response.json()) as Log[];
+    return json;
+}
 
-export { Log, getLogs };
+async function writeLog(user: string, tool: string, action: string, source: string): number {
+    const body = { infos: `${user},${tool},${source},${action}` };
+
+    const response = await fetch(endpoint, {
+        method: "post",
+        body: JSON.stringify(body),
+        headers: { "Content-Type": "application/json" },
+    });
+    await response.json();
+
+    return response.status;
+}
+
+export { Log, getLogFiles, getLogs, writeLog };